Release 1.6.0 on 2008-11-02.
I'm not sure at all about the original release date, but at least it's released in 2008.

This commit/release does not contain new maps because 1.5 release seems to accidentally have all the same set of maps as this release. Also the original 1.6 release did not contain any changes in SunBEAM.txt, but I have an (yet) unreleased version of SunBEAM.txt that states that the only changes between original releases 1.5 and 1.6 is some new maps (no map names specified). So as I were not able to define the correct set of maps in 1.5, this 1.6 release does not introduce the new maps it originally did - and I even cannot find out which maps those were.